# Reducing compute time for hexendec by a huge amount.