--- a/include/quicr/hex_endec.h
+++ b/include/quicr/hex_endec.h
@@ -2,15 +2,15 @@
 
 #include <quicr/quicr_name.h>
 
+#include <algorithm>
 #include <array>
+#include <bitset>
+#include <cassert>
 #include <cstdint>
 #include <iomanip>
 #include <limits>
 #include <sstream>
 #include <vector>
-#include <cassert>
-#include <algorithm>
-#include <bitset>
 
 namespace quicr {
 /**
@@ -65,9 +65,7 @@ public:
     static_assert(sizeof...(N) == sizeof...(UInt_ts),
                   "Number of values should match distribution of bits");
 
-    std::vector<uint8_t> distribution;
-    (distribution.push_back(N), ...);
-
+    std::vector<uint8_t> distribution{N...};
     return Encode(distribution, std::forward<UInt_ts>(values)...);
   }
   
@@ -78,41 +76,57 @@ public:
     static_assert(is_valid_uint<UInt_ts...>::value, "Arguments must all be unsigned integers");
     assert(distribution.size() == sizeof...(UInt_ts) && "Number of values should match distribution of bits");
 
-    std::vector<uint64_t> vals;
-    (vals.push_back(values), ...);
-
+    std::vector<uint64_t> vals{values...};
     return Encode(distribution, vals);
   }
 
-  static inline std::string Encode(const std::vector<uint8_t>& distribution, const std::vector<uint64_t>& values)
+  template<bool B = Size <= sizeof(uint64_t) * 8>
+  static inline typename std::enable_if<B, std::string>::type
+  Encode(const std::vector<uint8_t>& distribution, const std::vector<uint64_t>& values)
   {
     static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
     assert(distribution.size() == values.size() && "Number of values should match distribution of bits");
 
     std::ostringstream ss;
     ss << std::hex << "0x" << std::setfill('0');
-
-    std::bitset<Size> bits;
-    uint8_t remaining_bits = Size;
-    int i = 0;
+    
+    uint8_t current_distribution = 0;
+    uint64_t bits = 0;
     for (const auto& value : values)
     {
-      uint8_t dist = distribution.at(i++);
-      remaining_bits -= dist;
-      bits |= std::bitset<Size>(value) << remaining_bits;
+      const uint8_t dist = distribution.at(current_distribution++);
+      bits <<= dist;
+      bits |= (value & ~(~0x0ull << dist));
+    };
+    
+    ss << std::setw(Size / 4) << bits;
+
+    return ss.str();
+  }
+
+  template<bool B = Size <= sizeof(uint64_t) * 8>
+  static inline typename std::enable_if<!B, std::string>::type
+  Encode(const std::vector<uint8_t>& distribution, const std::vector<uint64_t>& values)
+  {
+    static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
+    assert(distribution.size() == values.size() && "Number of values should match distribution of bits");
+
+    std::ostringstream ss;
+    ss << std::hex << "0x" << std::setfill('0');
+    
+    uint8_t current_distribution = 0;
+    std::bitset<Size> bits;
+    for (const auto& value : values)
+    {
+      uint8_t dist = distribution.at(current_distribution++);
+      bits <<= dist;
+      bits |= std::bitset<Size>(dist >= sizeof(uint64_t) * 8 ? value : (value & ~(~0x0ull << dist)));
     };
 
-    if constexpr(Size > sizeof(uint64_t) * 8)
-    {
-      static const std::bitset<Size> bitset_divider(std::numeric_limits<uint64_t>::max());
-      ss << std::setw(Size / 8) << (bits >> (Size / 2) & bitset_divider).to_ullong();
-      ss << std::setw(Size / 8) << (bits & bitset_divider).to_ullong();
-    }
-    else
-    {
-      ss << std::setw(Size / 4) << bits.to_ullong();
-    }
-
+    static const std::bitset<Size> bitset_divider(std::numeric_limits<uint64_t>::max());
+    ss << std::setw(Size / 8) << (bits >> (Size / 2) & bitset_divider).to_ullong();
+    ss << std::setw(Size / 8) << (bits & bitset_divider).to_ullong();
+    
     return ss.str();
   }
 
@@ -134,12 +148,8 @@ public:
     static_assert(Size >= (N + ...), "Total bits cannot exceed specified size");
     static_assert(is_valid_uint<Uint_t>::value, "Type must be unsigned integer");
 
-    std::vector<uint8_t> distribution;
-    (distribution.push_back(N), ...);
-    
-    ValidateString(hex);
-
-    auto result = Decode(distribution, quicr::Name{hex});
+    std::vector<uint8_t> distribution{N...};
+    auto result = Decode(distribution, hex);
     std::array<Uint_t, sizeof...(N)> out;
     std::copy_n(std::make_move_iterator(result.begin()), sizeof...(N), out.begin());
 
@@ -153,55 +163,79 @@ public:
     static_assert(Size >= (N + ...), "Total bits cannot exceed specified size");
     static_assert(is_valid_uint<Uint_t>::value, "Type must be unsigned integer");
 
-    std::vector<uint8_t> distribution;
-    (distribution.push_back(N), ...);
-
-    auto result = Decode(distribution, name);
+    std::vector<uint8_t> distribution{N...};
+    auto result = Decode(distribution, name.to_hex());
     std::array<Uint_t, sizeof...(N)> out;
     std::copy_n(std::make_move_iterator(result.begin()), sizeof...(N), out.begin());
 
     return out;
   }
   
-  template<typename Uint_t = uint64_t>
-  static inline std::vector<Uint_t> Decode(const std::vector<uint8_t>& distribution, const std::string& hex)
+  template<typename Uint_t = uint64_t, bool B = Size <= sizeof(uint64_t) * 8>
+  static inline typename std::enable_if<B, std::vector<Uint_t>>::type
+  Decode(const std::vector<uint8_t>& distribution, const std::string& hex)
   {
-    ValidateString(hex);
-    return Decode<Uint_t>(distribution, quicr::Name{hex});
+    static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
+    static_assert(is_valid_uint<Uint_t>::value, "Type must be unsigned integer");
+
+    const auto dist_size = distribution.size();
+    std::vector<uint64_t> result(dist_size);
+    uint64_t bits = std::stoull(hex, nullptr, 16);
+    for (int i = dist_size - 1; i >= 0; --i)
+    {
+      const auto dist = distribution.at(i);
+      result[i] = bits & ~(~0x0ull << dist);
+      bits >>= dist;
+    }
+
+    return result;
+  }
+
+  template<typename Uint_t = uint64_t, bool B = Size <= sizeof(uint64_t) * 8>
+  static inline typename std::enable_if<!B, std::vector<Uint_t>>::type
+  Decode(const std::vector<uint8_t>& distribution, const std::string& hex)
+  {
+    static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
+    static_assert(is_valid_uint<Uint_t>::value, "Type must be unsigned integer");
+
+
+    std::string clean_hex = hex;
+    if (clean_hex.substr(0, 2) == "0x")
+      clean_hex.erase(0, 2);
+
+    constexpr uint8_t hex_length = Size / 4;
+    if (clean_hex.length() != hex_length)
+      throw std::runtime_error("Hex string value must be " +
+                              std::to_string(hex_length) + " characters (" +
+                              std::to_string(Size / 8) + " bytes)");
+
+    std::bitset<Size> bits;
+    const uint8_t section_length = hex_length / 2;
+    constexpr size_t sizeof_uint64_bits = sizeof(uint64_t) * 8;
+    constexpr size_t num_sections = Size / sizeof_uint64_bits;
+    for (size_t i = 0, j = 0; i < (section_length * num_sections); i += section_length)
+    {
+      const size_t midpoint = i + section_length;
+      std::string section = clean_hex.substr(i, midpoint);
+      bits |= std::bitset<Size>(std::stoull(section, nullptr, 16)) << (sizeof_uint64_bits * (num_sections - ++j));
+    }
+
+    const auto dist_size = distribution.size();
+    std::vector<uint64_t> result(dist_size);
+    for (size_t i = 0; i < dist_size; ++i)
+    {
+      const auto dist = distribution.at(i);
+      result[i] = (bits >> (Size - dist)).to_ullong();
+      bits <<= dist;
+    }
+
+    return result;
   }
   
   template<typename Uint_t = uint64_t>
   static inline std::vector<Uint_t> Decode(const std::vector<uint8_t>& distribution, const quicr::Name& name)
   {
-    static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
-    static_assert(is_valid_uint<Uint_t>::value, "Type must be unsigned integer");
-
-    constexpr uint8_t size_of_name = quicr::Name::size() * 8;
-    quicr::Name temp = name << (size_of_name - Size);
-    
-    std::vector<Uint_t> result(distribution.size());
-    for (size_t i = 0; i < distribution.size(); ++i)
-    {
-      auto dist = distribution.at(i);
-      result[i] = std::stoull((temp >> (size_of_name - dist)).to_hex(), nullptr, 16);
-      temp <<= dist;
-    };
-
-    return result;
-  }
-
-private:
-  static inline void ValidateString(const std::string& hex)
-  {
-    std::string clean_hex = hex;
-    auto found = clean_hex.substr(0, 2).find("0x");
-    if (found != std::string::npos)
-      clean_hex.erase(found, 2);
-
-    if (clean_hex.length() != Size / 4)
-      throw std::runtime_error("Hex string value must be " +
-                               std::to_string(Size / 4) + " characters (" +
-                               std::to_string(Size / 8) + " bytes)");
+    return Decode(distribution, name.to_hex());
   }
 };
 }