From some real world code, where we are on `main`:
```
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_HexEndecRealEncode               387 ns          387 ns      1783499
BM_HexEndecRealDecode              2001 ns         2001 ns       347007
```
Compared to this PR:
```
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_HexEndecRealEncode               355 ns          355 ns      1941952
BM_HexEndecRealDecode               534 ns          534 ns      1277745 
```
This mostly seems to have affected decode, but encode has certainly seen some benefits. Tried doing this by hand without bitset, but that actually resulted in higher compute times.