@GhostofCookie 2023-02-15
Here's some other tests that had been run just for comparison of a few more examples:
Before:
```
Run on (8 X 24.2411 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 4.82, 2.24, 1.57
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_HexEndecEncode4x32_to_128        362 ns          362 ns      1940956
BM_HexEndecDecode128_to_4x32       1391 ns         1391 ns       504014
BM_HexEndecEncode4x16_to_64         230 ns          230 ns      3053648
BM_HexEndecDecode64_to_4x16        1385 ns         1385 ns       500218
BM_HexEndecRealEncode               387 ns          387 ns      1783499
BM_HexEndecRealDecode              2001 ns         2001 ns       347007
```
After:
```
Run on (8 X 24.2406 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 1.24, 1.20, 1.22
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_HexEndecEncode4x32_to_128        341 ns          341 ns      2079712
BM_HexEndecDecode128_to_4x32        243 ns          243 ns      2896835
BM_HexEndecEncode4x16_to_64         171 ns          171 ns      4112881
BM_HexEndecDecode64_to_4x16        64.6 ns         64.6 ns     10728956
BM_HexEndecRealEncode               355 ns          355 ns      1941952
BM_HexEndecRealDecode               534 ns          534 ns      1277745