@GhostofCookie 2023-02-15
Final numbers:
```
Run on (8 X 24.2406 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 1.38, 1.50, 1.46
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_HexEndecEncode4x32_to_128        173 ns          173 ns      3878073
BM_HexEndecDecode128_to_4x32        190 ns          190 ns      3673673
BM_HexEndecEncode4x16_to_64         153 ns         51.2 ns     14628377
BM_HexEndecDecode64_to_4x16        32.2 ns         32.2 ns     21698093
BM_HexEndecRealEncode               187 ns          187 ns      3785666